feat: Added `checkout` menu
Added `git checkout` cli menu.
If there are no differences from the main, do nothing and exit.
If there are unstaged files, it asks the user to pick from them,
But if there only staged files, it asks the users if they want to
unstage any of them, then resets and checkouts out automatically.